+++ lib/nanoc/base/compilation/rule_context.rb
@@ -29,12 +29,13 @@ module Nanoc
       compiler = params[:compiler] or raise ArgumentError, "Required :compiler option is missing"
 
       super({
-        :rep     => rep,
-        :item    => rep.item,
-        :site    => compiler.site,
-        :config  => compiler.site.config,
-        :items   => compiler.site.items,
-        :layouts => compiler.site.layouts
+        :rep      => rep,
+        :item_rep => rep,
+        :item     => rep.item,
+        :site     => compiler.site,
+        :config   => compiler.site.config,
+        :items    => compiler.site.items,
+        :layouts  => compiler.site.layouts
       })
     end