+++ NEWS.md
@@ -1,5 +1,12 @@
 # nanoc news
 
+## 3.4.3 (???)
+
+Improvements:
+
+* Item reps are now accessible in a consistent way: in Rules and during
+  compilation, they can be accessed using both `@rep` and `@item_rep`
+
 ## 3.4.2 (2012-11-01)
 
 Fixes: